broadway: Restore guint32 type for memory size

Closes #3702

See merge request GNOME/gtk!3244

+++ gdk/broadway/gdkbroadway-server.c
@@ -235,7 +235,7 @@ static void
 parse_all_input (GdkBroadwayServer *server)
 {
   guint8 *p, *end;
-  size_t size;
+  guint32 size;
   BroadwayReply *reply;
 
   p = server->recv_buffer;
